refactor: move the running of subcommands into a submodule called subcommands
